Added missing prop: movies={movies} to my DirectorView and GenreView components
Finally realized I had failed to pass the prop: movies={movies} to my DirectorView component, and that's why the related-movies weren't rendering. Also added this prop to GenreView.

The content is rendering now, but still need to go back and address this warning:

Warning: Each child in a list should have a unique "key" prop. Check the render method of `DirectorView`